pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["troll_treasure", "tests"]

[project]
name = "troll_treasure"
version = "0.0.1"
requires-python = ">=3.6"
dependencies = ["art ~= 5.7.0", "pyYAML ~= 6.0.0"]

[project.optional-dependencies]
dev = ["pytest ~= 7.1.2"]

[project.scripts]
play = "troll_treasure.game:play"

